requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# Resume parsing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# NLP for skill extraction (optional - we use regex-based extraction)
# spacy==3.7.2
# Run: python -m spacy download en_core_web_sm

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# HTML templating
jinja2>=3.1.0

# CLI interface
click>=8.1.0
rich>=13.0.0

# Date handling
python-dateutil>=2.8.0
